[#27] Refactor: 일기 제목부분 제거 및 수정
