+++ src/main/java/umc/GrowIT/Server/domain/Diary.java
@@ -17,9 +17,8 @@ public class Diary extends BaseEntity {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(length = 50)
-    private String title;
-
+    @Lob // 필드를 TEXT로 매핑
+    @Column(nullable = false)
     private String content;
 
     @ManyToOne(fetch = FetchType.LAZY)
